feat: delete multiple selected suppliers

+++ app/Http/Livewire/Supplier/SupplierModule.php
@@ -7,9 +7,24 @@ use App\Models\Supplier;
 
 class SupplierModule extends Component
 {
-
-    //select for delet button
+    public $name = '';
+    public $phone = 8;
+    public $email;
+    public $address;
+    public $created_at;
+    public $updated_at;
     public $selectedSuppliers = [];
+    public $openModalConfirmDelete = false;
+
+    public $show = false;
+
+    public function openModalSupplierDelete() {
+        $this->show = true;
+    }
+
+    public function closeModalSupplierDelete() {
+        $this->show = false;
+    }
 
     public function render()
     {
@@ -22,9 +37,22 @@ class SupplierModule extends Component
 
     public function deleteSelectedSuppliers()
     {
+        // Verificando se algum fornecedor foi selecionado
+        if (empty($this->selectedSuppliers)) {
+            session()->flash('error', 'Selecione ao menos um registro!');
+            return;
+        }
+
         Supplier::whereIn('id', $this->selectedSuppliers)->delete();
+        
         $this->selectedSuppliers = [];
+        
         session()->flash('message', 'Fornecedores apagados!');
     }
 
+    // Método para atualizar a lista de fornecedores selecionados
+    public function updatedSelectedSuppliers()
+    {
+
+    }
 }
